requirements.txt
fastapi 
uvicorn 
websockets 
twilio 
torch
torchaudio
langchain-groq
groq
chatterbox-tts
pydub 
numpy
python-dotenv